backend/pyproject.toml
[project]
name = "podcast-recording"
version = "0.1.0"
description = "A podcast recording application with human host and AI companion"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "pydantic>=2.3.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.20",
    "openai>=1.3.0",
    "elevenlabs>=0.2.24",
    "assemblyai>=0.18.0",
    "httpx>=0.24.1",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
]

[tool.uv]
# uv-specific configurations can go here

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 